requirements.txt
python-dotenv>=0.21.0
deepgram-sdk>=3.0.0,<4.0.0
openai>=0.27.0
pyaudio>=0.2.11
websockets>=10.0